:recycle: refactor : 사용자 관련 영속화 객체의 변환을 도메인과 영속화 객체에서 분리하고 별도의 매퍼 파일로 관리하도록 변경
