+++ src/main/kotlin/org/example/codereviewstudy/infrastructure/persistence/post/mapper/PostPersistenceMapper.kt
@@ -1,8 +1,8 @@
 package org.example.codereviewstudy.infrastructure.persistence.post.mapper
 
 import org.example.codereviewstudy.domain.post.model.Post
-import org.example.codereviewstudy.domain.user.model.toJpaEntity
 import org.example.codereviewstudy.infrastructure.persistence.post.PostJpaEntity
+import org.example.codereviewstudy.infrastructure.persistence.user.mapper.toJpaEntity
 
 fun Post.toJpaEntity(): PostJpaEntity {
     return PostJpaEntity(
